// sarah.client/src/app/services/api/rules-service/model/promptRuleDto.ts
/**
 * Sarah Rules Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PromptRuleTimerDtoModel } from './promptRuleTimerDto';


export interface PromptRuleDtoModel { 
    id?: number;
    name?: string | null;
    guidance?: string | null;
    isEnabled?: boolean;
    sortOrder?: number;
    timer?: PromptRuleTimerDtoModel;
}